+++ CHANGELOG.md
@@ -8,6 +8,8 @@ CHANGELOG
   printf "\e[38;5;208mOption 1\e[m\nOption 2" | fzf --ansi
   printf "\e[38:5:208mOption 1\e[m\nOption 2" | fzf --ansi
   ```
+- Support `border-{up,down}` as the synonyms for `border-{top,bottom}` in
+  `--preview-window`
 
 0.32.1
 ------